Move TestCase to test_utils.py
