+++ sdc/tests/test_utils.py
@@ -140,3 +140,25 @@ def skip_sdc_jit(msg_or_func=None):
     wrapper = unittest.skipIf(sdc.config.config_pipeline_hpat_default, msg or "sdc pipeline not supported")
     # wrapper = lambda f: f  # disable skipping
     return wrapper(func) if func else wrapper
+
+
+class TestCase(unittest.TestCase):
+
+    def numba_jit(self, *args, **kwargs):
+        import numba
+        import warnings
+        if 'nopython' in kwargs:
+            warnings.warn('nopython is set to True and is ignored', RuntimeWarning)
+        if 'parallel' in kwargs:
+            warnings.warn('parallel is set to True and is ignored', RuntimeWarning)
+        kwargs.update({'nopython': True, 'parallel': True})
+        return numba.jit(*args, **kwargs)
+
+    def sdc_jit(self, *args, **kwargs):
+        return sdc.jit(*args, **kwargs)
+
+    def jit(self, *args, **kwargs):
+        if sdc.config.config_pipeline_hpat_default:
+            return self.sdc_jit(*args, **kwargs)
+        else:
+            return self.numba_jit(*args, **kwargs)
